requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
gitdb2==2.0.4             # via gitpython
gitpython==2.1.11
smmap2==2.0.4             # via gitdb2
spinners==0.0.23
